Misc matches and fixes (#2703)
* Match mDoExt_morf_c::getPlayMode

* Misc cleanup

* Fix actor cull spheres

* Match daPy_py_c::getLastSceneSwordAtUpTime, JUTGamePad::testTrigger

* Fix improper demangling of dBgS_SphChk::SetCallback

* Fix d_camera rangef weak func order

* Match more alink_wolf funcs

+++ src/d/actor/d_a_npc2.cpp
@@ -551,7 +551,7 @@ BOOL daBaseNpc_c::btkCtrl() {
 
 /* 8014F148-8014F1C8 149A88 0080+00 1/1 0/0 0/0 .text            checkEndAnm__11daBaseNpc_cFf */
 int daBaseNpc_c::checkEndAnm(f32 param_0) {
-    switch (mpModelMorf->getPlayMode()) {
+    switch ((u8)mpModelMorf->getPlayMode()) {
     case J3DFrameCtrl::EMode_LOOP:
         return mpModelMorf->isLoop();
     case J3DFrameCtrl::EMode_NONE:
